Modified to skip completed segments

+++ transmissioncoefficient.py
@@ -7,7 +7,7 @@ global INPUTTEMPLATE = \
 &cntrl
   irest     = 1,
   ntx       = 5,
-  ig        = -1,
+  ig        = RANDOMSEED,
   dt        = 0.002,
   nstlim    = NSTLIM,
   nscm      = 500,
@@ -61,6 +61,108 @@ class HP35Classifier(object):
 
 class FrameJob(object):
     def __init__(self, restartpath, parmpath, maindir, classifier, nsims=20):
+        self.restartpath = restartpath
+        self.parmpath = parmpath
+        self.maindir = maindir
+        self.classifier = classifier
+        self.nsims = nsims
+
+    def segment_is_complete(self, segdir):
+        '''
+        Check if the segment in segdir is complete. Return True if so, and 
+        return False otherwise
+        '''
+        # The last portion of segdir should be a string of digits that denotes
+        # the segment number
+        segnum = segdir.split('/')[-1]
+        
+        mdoutpath = os.path.join(segdir, "{:s}.out".format(segnum))
+
+        if os.path.exists(mdoutpath):
+            with open(mdoutpath, 'r') as mdout:
+                last_line = mdout.readlines()[-1]
+            if "Total wall time" in last_line:
+                return True
+            else:
+                return False
+        else:
+            return False
+
+    def run(self):
+        # run self.nsims simulations. Each is performed until it reaches some 
+        # state
+        for isim in xrange(self.nsims):
+            # simdir contains one simulation that starts from self.restartpath
+            # and continues until it reaches any state. simdir contains multiple
+            # sort segments; each segment is checked to see if the simulation
+            # has reached a state
+            simdir = os.path.join(maindir, "{:02d}".format(isim))
+            # endpoint_path contains a single ascii character denoting the final
+            # state of the simulations; if this file already exists, then the 
+            # simulation was completed in a previous submission of this script.
+            endpoint_path = os.path.join(simdir, 'endpoint')
+
+            # if the simulation was completed in a previous submission of this
+            # script, skip to the next simulation
+            if os.path.exists(endpoint_path):
+                continue
+
+            os.chdir(simdir)
+
+            # Think of the status as the "current" state of the simulation,
+            # though it works as a toggle -- once the simulation arrives in a
+            # state other than -1, status stays at that value. We exit the 
+            # following while loop once status is nonnegative
+            status = -1
+            iseg = 0 
+            
+            current_restart = self.restartpath
+            while status == -1:
+                # Prepare a directory for this segment
+                segdir = os.path.join(simdir, "{:04d}".format(iseg)) 
+                coordpath = os.path.join(segdir, '{:04d}.nc'.format(iseg))
+                newrestartpath = os.path.join(segdir, '{:04d}.rst'.format(iseg))
+
+                if not os.path.exists(segdir):
+                    os.makedirs(segdir)
+                os.chdir(segdir)
+
+                # Check if the segment is already completed. If not, run the 
+                # segment. If the segment is completed and we got to this part 
+                # of the script, we know also that the endnpoint file did not
+                # exist, implying that status must still be -1.  Therefore we
+                # do not need to update status.
+                if not self.segment_is_complete(segdir):
+
+                    inputstr = INPUTTEMPLATE.replace('RANDOMSEED', numpy.random.randint())\
+                                            .replace('NSTLIM', '50000')\
+                                            .replace('NTWR', '50000')
+
+                    with open('{:04d}.in'.format(iseg), 'w+') as mdin:
+                        mdin.write(inputstr)
+
+
+                    script = PMEMDSCRIPT.replace('DIR', segdir)\
+                                        .replace('RESTART', current_restart)\
+                                        .replace('TOPOLOGY', self.parmpath)\
+                                        .replace('MDOUT', coordpath)\
+                                        .replace('NEWRESTART', newrestartpath)
+                    with open('run.pmemd','w+') as f:
+                        f.write(script)
+
+                    # Run the simulation segment
+                    subprocess.Popen(['bash', 'run.pmemd']).wait()
+
+                    classifier = self.classifier(coordpath, self.parmpath)
+                    status = classifier.run()
+
+                current_restart = newrestartpath
+                iseg += 1
+
+                if iseg > 100:
+                    break
+            with open(endpoint_path, 'w+') as endpoint_file:
+                endpoint_file.write("{:d}".format(status))
 
 class SegmentJob(object):
     '''
@@ -88,7 +190,7 @@ class SegmentJob(object):
         self.segid = segid
         self.simdir = simdir
         self.parmpath = parmpath
-        self.indicatorarr = indicatorarr
+       sdf self.indicatorarr = indicatorarr
         self.analysisdir = analysisdir
         self.classifier = classifier
 
@@ -119,32 +221,81 @@ class SegmentJob(object):
                 raise ValueError
         return os.path.join(self.simdir, extension) 
 
-    def run(self):
-        # First, re-run the simulation
-        restartpath = self.get_restart_path()
+    def get_seed(self):
+        # find the path to the mdout file from the previous run of the simulation
+        if 'ff14sb.xray.1' in simdir or 'ff14sb.nmr.1' in simdir:
+            mdoutpath = os.path.join(self.simdir, 
+                                     "{:04d}".format(self.segid),
+                                     "{:04d}.out".format(self.segid))
+        else:
+            mdoutpath = os.path.join(self.simdir, 
+                                     "{:05d}".format(self.segid),
+                                     "{:05d}.out".format(self.segid))
+        # Scan the file for the line where the random seed is specified
+        with open(mdoutpath, 'r') as mdout:
+            for line in mdout:
+                if 'random seed' in line:
+                    break
+        seed = int(line.split()[8])
+        return seed
 
-        # run the simulation for 1 nanosecond, and save a new restart file every 10 ps
-        inputstr = INPUTTEMPLATE.replace('NSTLIM','500000').replace('NTWR', '-5000')
+    def restart_files_already_generated(self, restartdir):
+        '''
+        Return True if the restart files for this segment were already generated
+        and return False otherwise.
+        '''
+        mdoutpath = os.path.join(restartdir, 'mdout')
+        if os.path.exists(mdoutpath):
+            with open(mdoutpath, 'r') as mdout:
+                lastline = mdout.readlines()[-1]
+            if 'Total wall time' in lastline:
+                return True
+            else:
+                return False
+        else:
+            return False
+
+
+    def run(self):
+        # First, re-run the simulation so we have restart files at every 
+        # necessary time point
 
         # Restartdir will contain all the restart files that we need for starting the
         # transmission coefficient calculation simulations
         restartdir = os.path.join(self.analysisdir, 
                                   "{:05d}".format(self.segid),
                                   "restarts")
-        os.makedirs(restartdir)
+        if not os.path.exists(restartdir):
+            os.makedirs(restartdir)
+
         os.chdir(restartdir)
 
-        script = PMEMDSCRIPT.replace('DIR', restartdir)\
-                            .replace('RESTART', restartpath)\
-                            .replace('TOPOLOGY', self.parmpath)\
-                            .replace('MDOUT', 'mdout')\
-                            .replace('NEWRESTART', 'restart')
+        # Only do this next section if it was not already completed
+        if not self.restart_files_already_generated(restartdir):
+            # Get the path to the restart file that was originally used to start
+            # this simulation segment
+            restartpath = self.get_restart_path()
 
-        with open('run.pmemd', 'w+') as f:
-            f.write(script)
+            # To rerun the simulation, we need the random seed that was used the first time.
+            seed = self.get_seed()
 
-        # This should take between 5 and 10 minutes.
-        subprocess.Popen(['bash', 'run.pmemd']).wait()
+            # run the simulation for 1 nanosecond, and save a new restart file every 10 ps
+            inputstr = INPUTTEMPLATE.replace('NSTLIM','500000')\
+                                    .replace('NTWR', '-5000')\
+                                    .replace('RANDOMSEED', "{:d}".format(seed))
+
+
+            script = PMEMDSCRIPT.replace('DIR', restartdir)\
+                                .replace('RESTART', restartpath)\
+                                .replace('TOPOLOGY', self.parmpath)\
+                                .replace('MDOUT', 'mdout')\
+                                .replace('NEWRESTART', 'restart')
+
+            with open('run.pmemd', 'w+') as f:
+                f.write(script)
+
+            # This should take between 5 and 10 minutes.
+            subprocess.Popen(['bash', 'run.pmemd']).wait()
 
         # At this point, we should be finished saving all the restart files for the specified segment
         # Now it is a matter of looping over the segments for which we need to calculate the transmission coefficient
@@ -156,11 +307,8 @@ class SegmentJob(object):
                                      "{:05d}".format(self.segid), 
                                      "{:07d}".format(itp))
                job = FrameJob(restartfile, parmpath, jobdir, self.classifier, nsims=self.N)
+               job.run()
  
-
-        
-
-
 
 class TransmissionCalculation(object):
     '''
@@ -178,10 +326,31 @@ class TransmissionCalculation(object):
     simdir: (str) The path to the main simulation directory
     '''
 
-    def __init__(self, indicatorfile, parmpath, simdir):
+    def __init__(self, indicatorfile, parmpath, simdir, analysisdir):
 
         self.indicator = numpy.load(indicatorfile)
         self.parmpath = parmpath
         self.simdir = simdir
+        self.analysisdir = analysisdir
 
-     
+        # Generate a list of jobs
+        for iseg in xrange(1,10001):
+            indicators = self.indicator[100*i:100*(i+1)]
+
+            analysisdir = 
+            classifier = 
+
+            job = SegmentJob(iseg, self.simdir, self.parmpath, indicators,
+                             self.analysisdir, classifier) 
+            job.run()
+
+if __name__ == "__main__":
+    indicatorfile = '/gscratch3/lchong/ajd98/villin/transmission/indicatorfiles/ff14sb.xray.1.npy'
+    parmpath = '/gscratch3/lchong/ajd98/villin/ff14sb.xray.1/1_leap/VILLIN.parm7'
+    simdir = '/gscratch3/lchong/ajd98/villin/ff14sb.xray.1/'
+    analysisdir = '/gscratch3/lchong/ajd98/villin/transmission/ff14sb.xray.1'
+
+    TransmissionCalculation(indicatorfile,
+                            parmpath,
+                            simdir,
+                            analysisdir)
